Build system and code changes to support Nix builds
Modems are not as common nowadays, so this also includes a default
change to xtel.lignes, making the default modem a pty. You can create
it with e.g. socat, and point it to e.g. an application server running
over TCP or a UDS. This makes testing easier and hardware-free.

TODO: this commit also temporarily removes modem locking, as i have
not figured out why locking failures make xteld crash yet. This should
be avoided, and locking should be restored.